HL7 messages are not shown in web admin: Error: java.lang.Long cannot be cast to java.lang.Integer | Request method 'GET' not supported - TRUNK-3576 Author: ulakkaraju
git-svn-id: http://svn.openmrs.org/openmrs/trunk@28588 5bac5841-c719-aa4e-b3fe-cce5062f897a

+++ api/src/main/java/org/openmrs/hl7/db/HL7DAO.java
@@ -111,7 +111,7 @@ public interface HL7DAO {
 	 * @return count of HL7 items
 	 */
 	@SuppressWarnings("rawtypes")
-	public Integer countHL7s(Class clazz, Integer messageState, String query);
+	public Long countHL7s(Class clazz, Integer messageState, String query);
 	
 	/* HL7InArchive */
